requirements.txt
streamlit>=1.24
transformers>=4.30
torch>=2.0
python-docx
nltk
